split exam routes into 3 parts: exams, students, and questions
