--- routes/exams/middleware/index.js
@@ -1,61 +1,0 @@
-import {
-  checkIfExamIdIsValid,
-  doesCreateExamRequestHaveRequiredParams,
-  doesEditExamRequestHaveRequiredParams,
-  areCreateExamRequestParamsCorrectTypes,
-  isEditExamRequestParamCorrectType,
-} from './examValidation';
-import {
-  checkIfQuestionIdIsValid,
-  doesQuestionsArrayHaveObjElements,
-  doesQuestionsArrayElementsHaveRequiredParams,
-  doesAddQuestionRequestHaveRequiredParams,
-  doesEditQuestionRequestHaveRequiredParams,
-  areQuestionsArrayElementsParamsCorrectTypes,
-  areQuestionRequestParamsCorrectTypes,
-  doesQuestionsArrayElementsHaveCorrectParamsForType,
-  doesAddQuestionRequestHaveCorrectParamsForType,
-  doesEditQuestionRequestHaveCorrectParamsForType,
-} from './questionsValidation';
-import {
-  doesStudentsArrayHaveObjElements,
-  doesRequestContainValidStudentId,
-  doesStudentsArrayElementsHaveRequiredParams,
-  doesAddStudentRequestHaveRequiredParams,
-  doesEditStudentRequestHaveRequiredParams,
-  doesSaveExamResultsRequestHaveRequiredParams,
-  areStudentsArrayElementsParamsCorrectTypes,
-  areAddStudentRequestParamsCorrectTypes,
-  areEditStudentNameRequestParamsCorrectTypes,
-  areSaveExamResultsRequestParamsCorrectTypes,
-  doSaveExamResultsRequestArrayParamsHaveCorrectElements,
-} from './studentsValidation';
-
-export {
-  checkIfExamIdIsValid,
-  doesCreateExamRequestHaveRequiredParams,
-  doesEditExamRequestHaveRequiredParams,
-  areCreateExamRequestParamsCorrectTypes,
-  isEditExamRequestParamCorrectType,
-  checkIfQuestionIdIsValid,
-  doesQuestionsArrayHaveObjElements,
-  doesQuestionsArrayElementsHaveRequiredParams,
-  doesAddQuestionRequestHaveRequiredParams,
-  doesEditQuestionRequestHaveRequiredParams,
-  areQuestionsArrayElementsParamsCorrectTypes,
-  areQuestionRequestParamsCorrectTypes,
-  doesQuestionsArrayElementsHaveCorrectParamsForType,
-  doesAddQuestionRequestHaveCorrectParamsForType,
-  doesEditQuestionRequestHaveCorrectParamsForType,
-  doesStudentsArrayHaveObjElements,
-  doesRequestContainValidStudentId,
-  doesStudentsArrayElementsHaveRequiredParams,
-  doesAddStudentRequestHaveRequiredParams,
-  doesEditStudentRequestHaveRequiredParams,
-  doesSaveExamResultsRequestHaveRequiredParams,
-  areStudentsArrayElementsParamsCorrectTypes,
-  areAddStudentRequestParamsCorrectTypes,
-  areEditStudentNameRequestParamsCorrectTypes,
-  areSaveExamResultsRequestParamsCorrectTypes,
-  doSaveExamResultsRequestArrayParamsHaveCorrectElements,
-};
