+++ routes/exams/routes.js
@@ -1,45 +1,23 @@
 import express from 'express';
-
 import {
   checkIfExamIdIsValid,
   doesCreateExamRequestHaveRequiredParams,
   doesEditExamRequestHaveRequiredParams,
   areCreateExamRequestParamsCorrectTypes,
   isEditExamRequestParamCorrectType,
-  checkIfQuestionIdIsValid,
   doesQuestionsArrayHaveObjElements,
-  doesQuestionsArrayElementsHaveRequiredParams,
-  doesAddQuestionRequestHaveRequiredParams,
-  doesEditQuestionRequestHaveRequiredParams,
-  areQuestionsArrayElementsParamsCorrectTypes,
-  areQuestionRequestParamsCorrectTypes,
-  doesQuestionsArrayElementsHaveCorrectParamsForType,
-  doesAddQuestionRequestHaveCorrectParamsForType,
-  doesEditQuestionRequestHaveCorrectParamsForType,
   doesStudentsArrayHaveObjElements,
-  doesRequestContainValidStudentId,
+  doesQuestionsArrayElementsHaveRequiredParams,
   doesStudentsArrayElementsHaveRequiredParams,
-  doesAddStudentRequestHaveRequiredParams,
-  doesEditStudentRequestHaveRequiredParams,
-  doesSaveExamResultsRequestHaveRequiredParams,
+  areQuestionsArrayElementsParamsCorrectTypes,
   areStudentsArrayElementsParamsCorrectTypes,
-  areAddStudentRequestParamsCorrectTypes,
-  areEditStudentNameRequestParamsCorrectTypes,
-  areSaveExamResultsRequestParamsCorrectTypes,
-  doSaveExamResultsRequestArrayParamsHaveCorrectElements,
-} from './middleware/index';
+  doesQuestionsArrayElementsHaveCorrectParamsForType,
+} from './middleware';
 import {
   addExamToDatabase,
   getExam,
   editExamTitle,
   deleteExamFromDatabase,
-  addQuestion,
-  editQuestion,
-  deleteQuestion,
-  addStudent,
-  editStudentName,
-  deleteStudent,
-  saveExamResults,
 } from './controller';
 
 const examRoutes = express.Router();
@@ -69,56 +47,5 @@ examRoutes
     editExamTitle,
   )
   .delete(checkIfExamIdIsValid, deleteExamFromDatabase);
-
-examRoutes
-  .route('/questions/:examId')
-  .post(
-    checkIfExamIdIsValid,
-    doesAddQuestionRequestHaveRequiredParams,
-    areQuestionRequestParamsCorrectTypes,
-    doesAddQuestionRequestHaveCorrectParamsForType,
-    addQuestion,
-  )
-  .put(
-    checkIfExamIdIsValid,
-    checkIfQuestionIdIsValid,
-    doesEditQuestionRequestHaveRequiredParams,
-    areQuestionRequestParamsCorrectTypes,
-    doesEditQuestionRequestHaveCorrectParamsForType,
-    editQuestion,
-  )
-  .delete(checkIfExamIdIsValid, checkIfQuestionIdIsValid, deleteQuestion);
-
-examRoutes
-  .route('/students/:examId')
-  .post(
-    checkIfExamIdIsValid,
-    doesAddStudentRequestHaveRequiredParams,
-    areAddStudentRequestParamsCorrectTypes,
-    addStudent,
-  )
-  .put(
-    checkIfExamIdIsValid,
-    doesRequestContainValidStudentId,
-    doesEditStudentRequestHaveRequiredParams,
-    areEditStudentNameRequestParamsCorrectTypes,
-    editStudentName,
-  )
-  .delete(
-    checkIfExamIdIsValid,
-    doesRequestContainValidStudentId,
-    deleteStudent,
-  );
-
-examRoutes
-  .route('/students/save-exam-results/:examId')
-  .put(
-    checkIfExamIdIsValid,
-    doesSaveExamResultsRequestHaveRequiredParams,
-    doesRequestContainValidStudentId,
-    areSaveExamResultsRequestParamsCorrectTypes,
-    doSaveExamResultsRequestArrayParamsHaveCorrectElements,
-    saveExamResults,
-  );
 
 export default examRoutes;
