+++ routes/exams/students/middleware.js
@@ -5,23 +5,9 @@ import {
 } from '../../../common/helper';
 import {
   isStudentIdValid,
-  doesStudentHaveRequiredParams,
   areStudentParamsCorrectTypes,
   getExamFromDatabase,
 } from '../helper';
-
-function doesStudentsArrayHaveObjElements(req, res, next) {
-  const { students } = req.body;
-  const doesStudentsArrHaveObjElements = students.every((student) =>
-    isValueCorrectType(student, 'object'),
-  );
-  if (doesStudentsArrHaveObjElements) {
-    next();
-  } else {
-    res.status(400);
-    res.json({ error: 'students must be an array of objects' });
-  }
-}
 
 async function doesRequestContainValidStudentId(req, res, next) {
   const { examId } = req.params;
@@ -32,21 +18,6 @@ async function doesRequestContainValidStudentId(req, res, next) {
   } else {
     res.status(400);
     res.json({ error: 'invalid studentId' });
-  }
-}
-
-function doesStudentsArrayElementsHaveRequiredParams(req, res, next) {
-  const { students } = req.body;
-  const doesStudentsArrHaveRequiredParams = students.every((student) =>
-    doesStudentHaveRequiredParams(student),
-  );
-  if (doesStudentsArrHaveRequiredParams) {
-    next();
-  } else {
-    res.status(400);
-    res.json({
-      error: 'each student element must have a name property',
-    });
   }
 }
 
@@ -93,32 +64,6 @@ function doesSaveExamResultsRequestHaveRequiredParams(req, res, next) {
   } else {
     res.status(400);
     res.json({ error: errorMessage });
-  }
-}
-
-function areStudentsArrayElementsParamsCorrectTypes(req, res, next) {
-  const { students } = req.body;
-  const incorrectTypeParamErrs = [];
-  students.forEach((student) => {
-    const {
-      doesStudentHaveCorrectTypes,
-      incorrectTypeParamErrsArr,
-    } = areStudentParamsCorrectTypes(student);
-    if (!doesStudentHaveCorrectTypes) {
-      incorrectTypeParamErrs.push(...incorrectTypeParamErrsArr);
-    }
-    return student;
-  });
-  if (incorrectTypeParamErrs.length === 0) {
-    next();
-  } else {
-    const uniqueErrsSet = new Set(incorrectTypeParamErrs);
-    res.status(400);
-    res.json({
-      error: `students' param values must be correct types: ${createList([
-        ...uniqueErrsSet,
-      ])}`,
-    });
   }
 }
 
@@ -195,13 +140,10 @@ async function doSaveExamResultsRequestArrayParamsHaveCorrectElements(
 }
 
 export {
-  doesStudentsArrayHaveObjElements,
   doesRequestContainValidStudentId,
-  doesStudentsArrayElementsHaveRequiredParams,
   doesAddStudentRequestHaveRequiredParams,
   doesEditStudentRequestHaveRequiredParams,
   doesSaveExamResultsRequestHaveRequiredParams,
-  areStudentsArrayElementsParamsCorrectTypes,
   areAddStudentRequestParamsCorrectTypes,
   areEditStudentNameRequestParamsCorrectTypes,
   areSaveExamResultsRequestParamsCorrectTypes,
